Passing 4_7 and 4_8

+++ 4_8.py
@@ -5,17 +5,17 @@ import subprocess
 import lib.utils as utils
 from lib.timing_leak_attack import find_signature, config_browser
 
-def main():
+def main(args):
     # Init variables.
     os.environ['SERVER_TIMEOUT'] = "0.005"
     server_timeout = float(os.environ['SERVER_TIMEOUT'])
     print 'comparison time for each byte is', server_timeout, 'seconds'
 
-    # Start server.
-    command = 'python {}'.format(os.path.join('lib', 'timing_leak_server.py'))
-    with open(os.devnull, 'rw') as devnull:
-        p = subprocess.Popen(command, shell=True, stdout=devnull, stderr=devnull)
-    print 'server started'
+    if not args.no_server:
+        # Start server.
+        command = 'python {}'.format(os.path.join('lib', 'timing_leak_server.py'))
+        p = subprocess.Popen(command, shell=True)
+        print 'server started'
 
     # Create browser and configure.
     browser = config_browser()
@@ -25,17 +25,20 @@ def main():
 
     print 'signature is', signature
 
-    p.terminate()
+    # Stop server.
+    if not args.no_server:
+        p.terminate()
 
 if __name__ == '__main__':
     import argparse
     parser = argparse.ArgumentParser(
         description=('Break HMAC-SHA1 with a slightly less artificial timing leak'
                      ' - Challenge 32 (Set 4) of Matasano Crypto Challenge.'))
-
+    parser.add_argument('--no-server',
+        help='Don\'t run the timing leak server.', action='store_true')
     args = parser.parse_args()
 
     try:
-        main()
+        main(args)
     except KeyboardInterrupt:
         pass
